// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
type ContentRelationshipFieldWithData<
  TCustomType extends
    | readonly (prismic.CustomTypeModelFetchCustomTypeLevel1 | string)[]
    | readonly (prismic.CustomTypeModelFetchCustomTypeLevel2 | string)[],
  TLang extends string = string,
> = {
  [ID in Exclude<
    TCustomType[number],
    string
  >["id"]]: prismic.ContentRelationshipField<
    ID,
    TLang,
    PickContentRelationshipFieldData<
      Extract<TCustomType[number], { id: ID }>,
      Extract<prismic.Content.AllDocumentTypes, { type: ID }>["data"],
      TLang
    >
  >;
}[Exclude<TCustomType[number], string>["id"]];
// ...
export interface BlogPostDocumentDataAuthorsItem {
  /**
   * Author Info field in *Blog Post → authors*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.authors[].author_info
   * - **Documentation**: https://prismic.io/docs/fields/content-relationship
   */
  author_info: ContentRelationshipFieldWithData<
    [{ id: "author"; fields: ["author_name", "author_image", "author_bio"] }]
  >;
}

type BlogPostDocumentDataSlicesSlice = BreadcrumbsSlice | FaqSlice;

/**
 * Item in *Blog Post → Meta Keywords*
 */
export interface BlogPostDocumentDataMetaKeywordsItem {
  /**
   * Meta Keywords Text field in *Blog Post → Meta Keywords*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_keywords[].meta_keywords_text
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_keywords_text: prismic.KeyTextField;
}

/**
 * Item in *Blog Post → Meta Authors*
 */
export interface BlogPostDocumentDataMetaAuthorsItem {
  /**
   * Author Name field in *Blog Post → Meta Authors*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_authors[].author_name
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  author_name: prismic.KeyTextField;
}

/**
 * Content for Blog Post documents
 */
interface BlogPostDocumentData {
  /**
   * Category field in *Blog Post*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.category
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  category: prismic.RichTextField;

  /**
   * Publication Date field in *Blog Post*
   *
   * - **Field Type**: Date
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.publication_date
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/date
   */
  publication_date: prismic.DateField;

  /**
   * Blog Article Heading field in *Blog Post*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.blog_article_heading
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  blog_article_heading: prismic.RichTextField;

  /**
   * Article Main Image field in *Blog Post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.article_main_image
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  article_main_image: prismic.ImageField<never>;

  /**
   * Main Article Content field in *Blog Post*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.main_article_content
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  main_article_content: prismic.RichTextField;

  /**
   * authors field in *Blog Post*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.authors[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  authors: prismic.GroupField<Simplify<BlogPostDocumentDataAuthorsItem>>;

  /**
   * Slice Zone field in *Blog Post*
   *
   * - **Field Type**: Slice Zone
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.slices[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/slices
   */
  slices: prismic.SliceZone<BlogPostDocumentDataSlicesSlice> /**
   * Meta Title field in *Blog Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A title of the page used for social media and search engines
   * - **API ID Path**: blog_post.meta_title
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Blog Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: A brief summary of the page
   * - **API ID Path**: blog_post.meta_description
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *Blog Post*
   *
   * - **Field Type**: Image
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_image
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/image
   */
  meta_image: prismic.ImageField<never>;

  /**
   * Meta Image Alt Text field in *Blog Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_image_alt_text
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_image_alt_text: prismic.KeyTextField;

  /**
   * Meta Keywords field in *Blog Post*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_keywords[]
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  meta_keywords: prismic.GroupField<
    Simplify<BlogPostDocumentDataMetaKeywordsItem>
  >;

  /**
   * Meta URL field in *Blog Post*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_url
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  meta_url: prismic.KeyTextField;

  /**
   * Meta Authors field in *Blog Post*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: blog_post.meta_authors[]
   * - **Tab**: SEO & Metadata
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  meta_authors: prismic.GroupField<
    Simplify<BlogPostDocumentDataMetaAuthorsItem>
  >;
}

/**
 * Blog Post document from Prismic
 *
 * - **API ID**: `blog_post`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogPostDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<BlogPostDocumentData>,
    "blog_post",
    Lang
  >;
// ...
type FaqSliceVariation = FaqSliceDefault;

/**
 * Faq Shared Slice
 *
 * - **API ID**: `Faq`
 * - **Description**: *None*
 * - **Documentation**: https://prismic.io/docs/slices
 */
export type FaqSlice = prismic.SharedSlice<"Faq", FaqSliceVariation>;
// ...
type BreadcrumbsSliceVariation = BreadcrumbsSliceDefault;

/**
 * Breadcrumbs Shared Slice
 *
 * - **API ID**: `breadcrumbs`
 * - **Description**: Breadcrumbs
 * - **Documentation**: https://prismic.io/docs/slices
 */
export type BreadcrumbsSlice = prismic.SharedSlice<
  "breadcrumbs",
  BreadcrumbsSliceVariation
>;